+++ src/components/settings.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { Box, ToggleButton, ToggleButtonGroup } from "@mui/material";
 import { PopUp } from "../lib/popup";
 import { useSettings } from "./ContextProvider";
@@ -11,9 +11,7 @@ export default function Settings(props: {
 }) {
     const { open, setOpen } = props;
     const [settings, setSettings] = useSettings();
-    useEffect(() => {
-        localStorage.setItem("settings", JSON.stringify(settings));
-    }, [settings]);
+
     const settingItems: {
         title: string;
         action: (e: React.ChangeEvent<HTMLInputElement>) => void;