+++ src/pages/profile.tsx
@@ -49,8 +49,8 @@ export default function Profile() {
 
     useEffect(() => {
         if (isInteger(userId) && (!requestedUser || requestedUser.id !== userId)) {
-            api.profile
-                .userProfile({ userId })
+            api.users
+                .profile({ userId })
                 .then((res) => {
                     setRequestedUser(res.data);
                     setTitle(`${res.data.name} | Metahkg`);
@@ -120,7 +120,7 @@ export default function Profile() {
                         }}
                     >
                         <img
-                            src={`/api/profile/avatars/${requestedUser.id}`}
+                            src={`/api/users/avatars/${requestedUser.id}`}
                             alt="User avatar"
                             height={isSmallScreen ? 150 : 200}
                             width={isSmallScreen ? 150 : 200}