+++ src/types/conversation/thread.ts
@@ -1,16 +1,3 @@
-import { userType } from "../user";
-import { commentType } from "./comment";
+import { Thread } from "metahkg-api/dist/types/thread/thread";
 
-export type threadType = {
-    id: number;
-    pin?: commentType;
-    op: userType;
-    c: number;
-    vote: number;
-    slink?: string;
-    title: string;
-    category: number;
-    lastModified: string;
-    createdAt: string;
-    conversation: commentType[];
-};
+export type threadType = Thread;